change license, add savestate api, change init() api.
init now calls malloc to allocate enough space for the maximum events.
this change was done due to previously having to modify scheduler.h to set
the maximum number of events, which wasn't ideal.

while i try to avoid dynamic memory in all projects emulation related, i've since
given up on that idea, due to the fact that dynamic memory for flexible libraries
is a must, otherwise the code quality / useability suffers.

the savestate api was to reduce the code duplication across my many emulators.
this same api was first implimented in gb_apu.

Zlib -> MIT, i like to be credited for my work :)